refactor: split backtesting config into a directory
